requirements.txt
asgiref==3.7.2
Django==3.2.3
djangorestframework==3.14.0
djoser==2.1.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
Pillow==10.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
webcolors-1.13